+++ src/game/tests.rs
@@ -531,6 +531,7 @@ mod vintage_cube_seasoned_pyromancer;
 mod vintage_cube_sedgemoor_witch;
 mod vintage_cube_senseis_divining_top;
 mod vintage_cube_sentinel_nameless_city;
+mod vintage_cube_serra_paragon;
 mod vintage_cube_shelldock_isle;
 mod vintage_cube_sheltering_landscape;
 mod vintage_cube_sheoldreds_edict;